requirements.txt
# Core dependencies
pydantic>=1.10.0,<2.0.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
ffmpeg-python>=0.2.0

# ROS2 bag reading
rosbags>=0.9.0

# Image processing
opencv-python>=4.5.0

# Utilities
tqdm>=4.60.0

# ROS2 message handling (if not available via system packages)
rosidl-runtime-py
